requirements.txt
Pillow==2.6.1
PyBluez==0.18
RPi.GPIO==0.6.2
RTIMULib==7.2.1
argparse==1.2.1
chardet==2.3.0
colorama==0.3.2
gdata==2.0.18
google-api-python-client==1.5.0
gpiozero==1.1.0
html5lib==0.999
httplib2==0.9.2
lxkeymap==0.1
mcpi==0.1.1
ndg-httpsclient==0.3.2
numpy==1.8.2
oauth2client==2.0.1
picamera==1.10
pifacecommon==4.2.1
pifacedigitalio==3.1.0
pyOpenSSL==0.13.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycrypto==2.6.1
pygame==1.9.2a0
pygobject==3.14.0
pyserial==2.6
python-dateutil==2.5.3
python-gcm==0.4
python-utils==1.6.2
pytz==2016.4
requests==2.4.3
rsa==3.4
sense-hat==2.1.0
simplejson==3.8.2
six==1.10.0
spidev==3.0
tlslite==0.4.9
tzlocal==1.2.2
uritemplate==0.6
urllib3==1.9.1
wheel==0.24.0
wsgiref==0.1.2
